Better builtins!
Allow masking of builtins, such that if a module defines a variable of the same
name as a builtin, this variable takes precedence over the builtin. This affords
future compatibility when adding new builtins.

Also allow builtins to be defined as variables, not just constants, so that we
can have lazily-computed builtins.